+++ app/src/main/java/dev/kdrag0n/android12ext/core/xposed/hooks/SystemUIHooks.kt
@@ -61,7 +61,7 @@ class SystemUIHooks(
 
     fun applyThemeOverlayController(
         isGoogle: Boolean,
-        chromaMultiplier: Float,
+        chromaMultiplier: Double,
         multiColor: Boolean,
         accurateShades: Boolean,
         colorOverride: Int?,
