+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/Cam16.kt
@@ -3,14 +3,14 @@ package dev.kdrag0n.android12ext.monet.colors
 import dev.kdrag0n.android12ext.monet.colors.google.Cam
 
 data class Cam16(
-    override val L: Float,
-    override val C: Float,
-    override val h: Float,
+    override val L: Double,
+    override val C: Double,
+    override val h: Double,
 ) : Lch {
     override fun toLinearSrgb() = toSrgb().toLinearSrgb()
 
     fun toSrgb(): Srgb {
-        val cam = Cam.fromJch(L, C, h)
+        val cam = Cam.fromJch(L.toFloat(), C.toFloat(), h.toFloat())
         return Srgb(cam.viewedInSrgb())
     }
 
@@ -18,9 +18,9 @@ data class Cam16(
         fun Srgb.toCam16(): Cam16 {
             val cam = Cam.fromInt(quantize8())
             return Cam16(
-                L = cam.j,
-                C = cam.chroma,
-                h = cam.hue,
+                L = cam.j.toDouble(),
+                C = cam.chroma.toDouble(),
+                h = cam.hue.toDouble(),
             )
         }
     }
