+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/CieXyz.kt
@@ -1,24 +1,24 @@
 package dev.kdrag0n.android12ext.monet.colors
 
 data class CieXyz(
-    val x: Float,
-    val y: Float,
-    val z: Float,
+    val x: Double,
+    val y: Double,
+    val z: Double,
 ) : Color {
     override fun toLinearSrgb(): LinearSrgb {
         return LinearSrgb(
-            r = +3.2404542f * x + -1.5371385f * y + -0.4985314f * z,
-            g = -0.9692660f * x + +1.8760108f * y + +0.0415560f * z,
-            b = +0.0556434f * x + -0.2040259f * y + +1.0572252f * z,
+            r = +3.2404542 * x + -1.5371385 * y + -0.4985314 * z,
+            g = -0.9692660 * x + +1.8760108 * y + +0.0415560 * z,
+            b = +0.0556434 * x + -0.2040259 * y + +1.0572252 * z,
         )
     }
 
     companion object {
         fun LinearSrgb.toCieXyz(): CieXyz {
             return CieXyz(
-                x = 0.4124564f * r + 0.3575761f * g + 0.1804375f * b,
-                y = 0.2126729f * r + 0.7151522f * g + 0.0721750f * b,
-                z = 0.0193339f * r + 0.1191920f * g + 0.9503041f * b,
+                x = 0.4124564 * r + 0.3575761 * g + 0.1804375 * b,
+                y = 0.2126729 * r + 0.7151522 * g + 0.0721750 * b,
+                z = 0.0193339 * r + 0.1191920 * g + 0.9503041 * b,
             )
         }
     }
