+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/Srlch2.kt
@@ -6,9 +6,9 @@ import dev.kdrag0n.android12ext.monet.colors.Lch.Companion.calcLchC
 import dev.kdrag0n.android12ext.monet.colors.Lch.Companion.calcLchH
 
 data class Srlch2(
-    override val L: Float,
-    override val C: Float,
-    override val h: Float,
+    override val L: Double,
+    override val C: Double,
+    override val h: Double,
 ) : Lch {
     override fun toLinearSrgb() = toSrlab2().toLinearSrgb()
